+++ src/app/routes/transactions/placeOrder.ts
@@ -331,6 +331,7 @@ placeOrderTransactionsRouter.post(
     async (req, res, next) => {
         try {
             const authorizeObject = {
+                typeOf: <'Mvtk'>'Mvtk',
                 // tslint:disable-next-line:no-magic-numbers
                 price: parseInt(req.body.price, 10),
                 transactionId: req.params.transactionId,