+++ CHANGELOG.md
@@ -9,6 +9,7 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/).
 
 ### Changed
 - マルチトークン発行者に対応。
+- 取引結果にポストアクションを定義。
 
 ### Deprecated