+++ CHANGELOG.md
@@ -10,6 +10,7 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/).
 ### Changed
 - マルチトークン発行者に対応。
 - 取引結果にポストアクションを定義。
+- CORS設定調整。
 
 ### Deprecated