+++ CHANGELOG.md
@@ -11,6 +11,7 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/).
 - マルチトークン発行者に対応。
 - アクションと取引に対して潜在アクション属性を定義。
 - CORS設定調整。
+- 承認アクションのobjectに型を定義し、purposeを取引型に変更。
 
 ### Deprecated